+++ plugins/jaeger/src/components/panel/details/SpanTag.tsx
@@ -1,4 +1,5 @@
-import React from 'react';
+import React, { useState } from 'react';
+import { Tooltip } from '@patternfly/react-core';
 
 import { IKeyValue } from '../../../utils/interfaces';
 
@@ -7,12 +8,28 @@ interface ISpanTagProps {
 }
 
 const SpanTag: React.FunctionComponent<ISpanTagProps> = ({ tag }: ISpanTagProps) => {
+  const [visible, setVisible] = useState<boolean>(false);
+
+  const copy = (): void => {
+    if (navigator.clipboard) {
+      setVisible(true);
+
+      navigator.clipboard.writeText(`${tag.key}=${tag.value}`);
+
+      setTimeout(() => {
+        setVisible(false);
+      }, 1500);
+    }
+  };
+
   return (
-    <div className="pf-c-chip pf-u-ml-sm pf-u-mb-sm" style={{ maxWidth: '100%' }}>
-      <span className="pf-c-chip__text" style={{ maxWidth: '100%' }}>
-        {tag.key}: {tag.value}
-      </span>
-    </div>
+    <Tooltip content={<div>copied</div>} isVisible={visible} trigger="manual">
+      <div className="pf-c-chip pf-u-ml-sm pf-u-mb-sm" style={{ cursor: 'pointer', maxWidth: '100%' }} onClick={copy}>
+        <span className="pf-c-chip__text" style={{ maxWidth: '100%' }}>
+          {tag.key}: {tag.value}
+        </span>
+      </div>
+    </Tooltip>
   );
 };